requirements.txt
requests==2.21.0
cloudant==2.12.0
python-dotenv==0.10.1

